Split Assert.AreEqual.Numerics.cs into per-numeric-type partial files
Fixes #9801

Split the 570-line Assert.AreEqual.Numerics.cs into a trimmed shared-helpers
file plus four per-type partial files (Float, Double, Decimal, Long), each
under 300 lines. No public API or behavior changes.
